Refactoring SuperAdminContext. Making useSuperadminContext hook to save all functions and make SuperAdminContext take care only about saving and sharing values
