+++ src/context/SuperAdminContext/index.ts
@@ -1,0 +1,1 @@
+export { SuperAdminProvider, useSuperAdminFunction } from "./SuperAdminContext";
